move HGCalImagingAlgo to standard packages
