+++ RecoLocalCalo/HGCalRecProducers/python/imagingClusterHGCal_cfi.py
